+++ AlgoHero/src/algo3c1g2/modelo/nota/Sol.java
@@ -5,7 +5,7 @@ import algo3c1g2.modelo.figura.Figura;
 public class Sol extends Nota {
 
 	//Frecuencia caracteristica del tono Sol en la octava 3
-	private final static float FRECUENCIA_SOL = 67;
+	private final static int FRECUENCIA_SOL = 67;
 	private final static int CUERDA = 4;
 	
 	public Sol(){
